Added test for constants and updated conditional compilation for constants.
Updated constant.cpp use NotPresent for those constants that are known but do
not exist on the platform where this module is compiled on.

New test file loops through exported constants and complains if a constant is
exported that is not known in constant.cpp.

Currently SERVICE_ACCEPT_DEVICEEVENT is exported but not defined in
constant.cpp.