Fixes bug in DllMain()
A bug in DllMain() makes the switch statement only work for
dwMajorVersion up to version 6. The effect is that gdwControlsAccepted
is set to zero (the default value) for dwMajorVersion greater than 6.
meaning that the SERVICE_ACCEPT_STOP, SERVICE_ACCEPT_PAUSE_CONTINUE, and
SERVICE_ACCEPT_SHUTDOWN bits will not be set. This makes it impossible
to stop, pause and shutdown a service from the "Services" app in Windows
10. Also the "net stop" command cannot stop the service.